+++ tests/test_device_specs.py
@@ -1,3 +1,4 @@
+import json
 from uuid import uuid4
 
 import pytest
@@ -5,7 +6,6 @@ import pytest
 from sdk import SDK
 
 
-@pytest.mark.only
 class TestDeviceSpecs:
     @pytest.fixture(autouse=True)
     def init_sdk(self, start_mock_request):
@@ -14,6 +14,7 @@ class TestDeviceSpecs:
         )
 
     def test_get_device_specs_success(self):
-        device_specs_list = self.sdk.get_device_specs_list()
-        assert len(device_specs_list) == 1
-        assert type(device_specs_list[0].specs) == dict
+        device_specs_dict = self.sdk.get_device_specs_dict()
+        assert type(device_specs_dict) == dict
+        specs = device_specs_dict["FRESNEL"]
+        json.loads(specs)